nuera-ai/requirements.txt
# requirements.txt
streamlit
sentence-transformers
faiss-cpu
torch
transformers
accelerate
